+++ CHANGELOG.md
@@ -5,3 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+
+### Added
+
+- PCA9554 driver implementation which implements Zephyr's GPIO API. This driver
+  does not support interrupt functionality. This driver was tested with
+  `nrf-sdk v2.1.0.`.
+- Sample which shows basic use and configuration.